main,pkg: remove pkg, all in main
